Adding ArtefactType::postcommit_hook function
Bug 1401210. In order for watchlists to properly handle blog
blocks, we need to give the Blog artefact a chance to update
the view_artefacts table before the "saveartefact" event gets
fired at the bottom of ArtefactType::commit().

So I'm adding a postcommit_hook() method which gets called in
the middle of commit(), and shifting ArtefactType's commit()
method into there.

Change-Id: I23986597fd411fa8f93da7cdbfa8d0b577de66fc