Make Publish/Unpublishing a blog entry, fire saveartefact event
Bug 1401210. In order for a newly published blog entry (using the
"publish" button) to be able to trigger a watchlist notification,
it needs to fire a saveartefact event, and update the view_artefact
table, the same as the standard route for editing/creating a blog
entry. So the best way to do that is to use the same commit() method.

Change-Id: I35b93d6f7415ae69ed8eede3e10306b712ea1b68