Behat Test: Tinymce Editor (Bug 1447449)
The description element cannot be interacted with still, we can interact
with a TinyMCE field via:
 And I fill in "text" in WYSIWYG editor "editview_description_ifr"

But what we are needing is the ability for behat to work out that
the <textarea> is hidden and so to cehck if it has a TinyMCE editor in
play and use that instead.

Instead of fixing the behat test we should fix behat itself.

Change-Id: I06f57bea875394d14f51ca1baa0ffe1ed012b791