Fixing the layout images being broken by locale setting
Bug 1468144

If your php site locale is one that uses comma as decimal seperator,
eg in Dutch 12,5 is the same as 12.5 in English then the layout thumbs
break.

Need to change , back to . for it to work.

The best way to do this would be to set the locale to english for the
layout creation and then set it back again but we can't be sure which
english is available so I've just done a str_replace of , to .

Change-Id: I807eda47e76d3654137681bded26164e535eb9eb
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>