Removing obsolete top submit button from profile (Bug #1465462)
This top submit button was used before we split the edit profile
page into tabs. The old tab JS used to remove it from the page.
The new Bootstrap JS does not remove it, and this causes problems
for Behat. But since we no longer support non-JS operation, it
makes more sense to remove it.

Change-Id: I38c5e451d018e47dc0704fe61892632431ec0c03