Updating the "random clicking" test (Bug 1465462)
With the Behat changes, the activity type filter and the
"mark as read" buttons are no longer present in your inbox
until after you've received some messages.

So, those would be better moved to a more focused inbox test,
rather than being in "random clicking".

I've added some additional random clicks in the inbox screen
to make up for it. ;)

Change-Id: I4388320b95b1e7d508a04367afca283de167e6d3