Bug 1468156: Change plugin type of plugin multirecipientnotification to module
To change the plugin type from artefact to module, the code is moved and the
superclass reference is changed to PluginModule. Obsolete classes for the
former unused artefact types are deleted. The references to the old script
paths are changed. Plugin Tables are renamed and references updated.
The new plugin mustn't be installed when the old plugin is still active, but on
upgrade, the database tables are renamed and references in artefact_installed
and artefact_event_subscription and the triggers are changed.

Change-Id: Id5f1198b780ffb92181f7d60785aad9ffe2259da
Signed-off-by: Tobias Zeuch <tobias.zeuch@rwth-aachen.de>