Remove the index name from elasticsearch JSON data (Bug 1457709)
If we remove the index name from the JSON data, then Elasticsearch
will just use the index name in the URL path, which we're already
setting correctly.

Doing it this way, we don't have to rely on the Elasticsearch server
setting the "rest.action.multi.allow_explicit_index: false" flag.

Change-Id: Ib6194b89e68d539a7959b9e12ec1ac9e3528dd27