Increase the size of the Leap2a entrycontent column (Bug 1474659)
In MySQL, the default "text" column size is 65kB. Because this column
can hold the serialized text from several other columns, it needs to
be larger than that. Setting it to "big" generates a "longtext", which
is 4GB.

This change has no effect in Postgres. Postgres has only one text
size (unlimited). The XMLDB generator is smart enough that it
recognizes this and doesn't run any SQL in this case.

Change-Id: Ic26f0eb3d3c3181fdcaf8729373ca87b6ecc9c8d