Straightening out the language string situation in the JS calendar (bug 1455137)
The main thing this fixes, is that it causes the JS calendar to use ONLY the
new lang strings pieform_calendar_dateformat & pieform_calendar_timeformat,
for its formatting. These strings are in strftime() format, and are converted
into the JQuery Datepicker/Timepicker format by PHP (if we change to a different
calendar plugin in the future, we can change the PHP conversion function).

This fixes the bug where translators could break the calendar by changing the
JS format strings so they no longer matched the PHP format strings.

As a side effect of this change, I've also made it so that the help strings
are derived from the format string, and are translatable.

behatnotneeded: Date field is already tested by resume_page.feature

Change-Id: I152962cc27f36f93b51ee0327ed942fb089f7f5b