Bug 1511536: problem with deleting blocks when edit page first loads
The problem is due to the scope of the blockinstanceId variable being
outside of the click handler function, which inadvertently makes it
into a sort of static variable shared by all of the delete buttons.

So the solution is to use a locally-scoped blockinstanceId button
inside the click handler method.

Change-Id: I3e8ffac31cc869c545e85c5f53c3e710e2e71243
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>
Signed-off-by: Aaron Wells <aaronw@catalyst.net.nz>
(cherry picked from commit 52770f8baf2297d9cbf5506e2696b346fae8f419)