"Tagged journal entries" block shouldn't grant access to whole journal
Bug 1521818. Making the "Tagged journal entries" block act more like
a collection of "Journal entry" blocks. So, it doesn't add the parent
blog to view_artefacts, only the specific blog entries that are
displayed in the block.

Also removing the title of the parent blog (and the link to it) from
the list of blog entries, like the "Journal entry" block, which
doesn't display the title of the containing journal.

Note the viewer may still have access to the whole blog, if the blog
is also shared on the same page via a "blog" or "recent journal entries"
block.

Change-Id: I33fc7e58b964c03bc8003f1de81a4bf58b6079b7
(cherry picked from commit ada12dba53c2da3596cdf51708ea5d666b60546e)