auth/imap: Autocreating users; email domain filter
Bug 1523644. This change adds an "autocreate users"
option to auth/imap, and it allows you to add a domain
name filter to check that the user's email address is
@ a particular domain.

Taken together, these changes make it possible to use
auth/imap with Google Apps for Business, and other
similar services.

behatnotneeded: Requires an IMAP server to test

Change-Id: Ic1892e4df652e3997d21897cc03bfb9cf1e075cf