Removing obsolete "disablelogin" setting
Bug 1526076: I believe the initial intent was that
Mahara core, and/or each plugin, could add a value
to its version.php file indicating "disablelogin"
true or false. And in this way, an upgrade could
indicate whether it was a small enough upgrade that
users did not need to log out for it.

However, in practice this is not practical because
we don't know what version of Mahara the user is
upgrading from, and that is what determines whether
or not it's a "stable" upgrade.

Additionally, the core disablelogin has been set to true
for the past 7 years, and the plugin disablelogin
setting no longer has any effect.

Removing disablelogin should hopefully make our
maze of init.php auth_setup() login stuff a little
bit easier to follow.

behatnotneeded: Covered by existing tests

Change-Id: I5f8a2b4faa95b9225bb926de6a54a622ea1a9618