Bug 1438894: Moving the profile introduction text to description column
Currently all the artefacts that save html/tinymce data do so in the
description field - except the internal profile introduction field.

Seen as we are already doing special handling of this plugin we might
as well save the html/tinymce data into the 'description' column of
the db for consistency sake.

behatnotneeded - existing tests should suffice

Change-Id: I68da79f1c9423e19218162d8315008134251c31f
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>