Bug 1544424: Only load block.js when it's actually needed
This prevents an infinite loop caused by an error in a ajax block
causing error.tpl to be rendered, which causes block.js to get loaded,
which causes a new ajax request for all the ajax blocks, which causes
error.tpl to be rendered, etc.

Also moving this JS file out of the theme directory and into the /js
directory, because it's core, not theme.

behatnotneeded: Can't test error state via Behat

Change-Id: Idc83002dbf5cb8af36d9d33e916da2bdfa5a308f