Bug 1545380: Profile images on user list broken for staff
When 'Staff report access' is set to 'no'. This is due to the column
being built from both the 'select' and 'icon' part of the $cols array.

So need to pass in a mostly blank 'select' column info rather than
unset it.

Also fixed up the missing right side line on button when viewing page as
staff with 'Staff report access' set to 'yes'

behatnotneeded

Change-Id: I97d77f0dc136967efa16945114bcd990e49754d0
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>