Allow byteranges all the time
Bug 1397131: iOS needs byte ranges for media files.
We have some very old code that prohibits byte ranges
when cacheing is disabled, which in turn is inherited
from some very old code in Moodle. But it doesn't
appear to be actually necessary to *ever* prohibit
byte ranges.

behatnotneeded: Can't be tested in Behat

Change-Id: Id3bba20305204f87cee50ec664893f53cfb2337e