Bug 1594579: Allowing a group to be duplicated
This patch does the copying of:

- The group settings - and renames the group to a new unique name
- Views/collections - the same way we copy them now so that copied files
  end up in 'viewfiles' folder.
- The group homepage by copying the old homepage like a normal
page then move the blocks from it to the new homepage and then deleting the
duplicated homepage.
- Adding copy buttons to the admin/groups list, the my groups list,
and group homepage.

Still to do:

Make sure we don't duplicate more than one copy of a journal if the
journal is within more than one page/block

behatnotneeded - behat file to come later

Change-Id: Ibd74d1bdf76c3488d4af3735dba9c7b5f773e2f7
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>