Bug 1598990: Verify the import format from an external source is correct.
Check the format of the record in the import_queue table when processing
the queue. Ensure it always defaults to a valid format that is installed
and enabled. If it is not valid, throw an error.

Clean up the call PluginImport::create_new_queue() by removing the
non-existant 'plugin' field.

Also, wrap the importer in a try/catch so the core_import_process_queue
cron process can contiue if there is an error.

And, validate that the keypair_PEM exists when checking openssl_open().

behatnotneeded

Change-Id: Icde78f862ca0e120b67393f4a5abf464695f5104