Bug 1580399: Stop users logging in to suspended/expired institutions
Moving the code from LiveUser->login() to
ensure_user_account_is_active() so that internal and external logins
can use the same code. This means the check now will fall after
LiveUser->authenticate() so a user's lastlogin values will be updated.
but that should be ok as the login was successful, it's just they
can't go any further as their institution is not active.

behatnotneeded

Change-Id: Ie78a60978d5936f78af5a962ca3efdcdee148b93
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>