Bug 1605127: Method signature mismatch in Activity::update_url()
The abstract Activity class defines the function update_url()
with no parameters, but the two subclasses that override it
define it with one parameter. This will cause problems in PHP 7.

It's always called with one parameter, (even in the Activity
class itself) so the best option is to add one parameter to the
implementation in Activity.

Change-Id: I810061ed6f8c55101327e2e907bb68ebf9870380
behatnotneeded: Covered by existing tests
(cherry picked from commit b469030b12e80147d406fc00b75afe024a329402)