Bug 1606094: Stop sending quota notifications to everyone
This patch changes that bad left (outer) join to an inner
join, so that we only check for notifications for users
who are in the affected institution.

As a further safeguard, it also removes the line that
compared each user's file usage to the institution's
new quota. Now, we'll instead compare each user's
file usage to that user's current quota (which, for
members of the institution, will be the institution's
new quota.)

Change-Id: I4db6068221d1238c8a755434e78e6cc006a84aa8
behatnotneeded: Test to come later
(cherry picked from commit 2c262cdf8d3ddb8c9c66565c084b67e2f24825e6)