Bug 1612451: Graceful handling of invalid embedded artefact id
Adds a try/catch block so that the code doesn't crash if
you've put in an invalid image ID in an embedded image tag.
Also tweaks the embed code so that it will skip invalid
images but continue processing the rest of the text block.

behatnotneeded: Covered by existing tests

Change-Id: I573adc7d29a0c1a4289f1b0a0b402d308a739be9