Bug 1626805: Allow the view constructor to fix missing layout info
Sometimes in MySql if many users are copying a page at once the
view_rows_columns information can get deleted.

As we can work out what the values should have been from other tables
we can regenerate it and add it back in.

If the view doesn't have a layout set it hasn't changed from when
first been made so we can make it the same as view type's template

To test: see bug report

behatnotneeded

Change-Id: I81c8e33dc1bc4121c1df97e1acbc981ad969271c
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>