Bug 1629144: Allow self copying of an annotation to duplicate artefact
Because unlike image/note blocks where we want the blocks to reference
the same underlying artefact here we want to create a new artefact
otherwise updating the block on one page will update the artefact for
all the related blocks.

To do this we need to add a new copytype called 'fullinclself' that
means make a full copy of blockinstance including when doing a self copy

Also we need to allow for the self copying to ignore the
annotationfeedback artefact type as we dont' want to copt the feedback.

behatnotneeded

Change-Id: Icc9e375e3180ff3f25d84ba9695926d5b24bb17d
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>
(cherry picked from commit a9360f18ba59ba48358add888cf83a036556f076)