Bug 1304146: Delete remote user when delete institution
If a user has manually assigned a remote authorization method
from an institution and he is not a member of that institution
when we try to delete the institution there will still be
a record in auth_remote_user table.

When the institution is deleted, so is the authorization method,
and so should be the remote user record.

behatnotneeded

Change-Id: I91dcadc5fbfddc8941c60adc78f304996591b3b8
(cherry picked from commit 933fc2666becd5b251610e1b3ffcf6a952e9c368)