Bug 1673295 - add behat test for "Some HTML" block
Replaced the html file with a new one that doesn't contain stylesheets.
Changed the test to check the content displayed in the block doesn't contain
html tags and that an 'a' tag is recognised as a link. Now links to the
Mahara manual.

NB: Lorempixel pics have Creative Commons license.

Change-Id: I2d4a5f2b260bdc970a87c4acbf5ac63350e8c535