Bug 1667539: "Institutions" in 2nd level nav to "Settings"
to avoid having "Institutions" shown up twice,
once in 1st level nav and then in 2nd level nav
again. For all other items in the 2nd level nav
we say what we are displaying just not for the
"Settings" page as the institutions as listed
differently than on the other pages.

Change-Id: I9b9248226fce76e485c8a898c04fe4eb2bf5e7e8