Bug 1681961: Ignore link to first page of collection on no SE page
If we create a SmartEvidence collection and add a page to it then
turn SmartEvidence off for the institution and try to go to the matrix
page, we get an error page with link to first view in collection.

If the collection has no pages though the matrix error page has a link
going nowhere.

behatnotneeded

Change-Id: I168bb2a9d30d7d45e6d54385458bc5e8012c0bcb
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>
(cherry picked from commit 799e06cf85346d439c9b1bac24c348f41accca7e)