Bug 1680640: Allow multiple submit buttons from one pieform element
Calling it multisubmit where we pass in an arrays of options/values so
that it sets up a bunch of 'submit' buttons that all flow thru the
same [formname]_submit function and be handled there.

This allows the submit buttons to 'live together' on the page and in
the code.

behatnotneeded

Change-Id: I16864d46ff3ea6bc0cb89e7769ba4caa5c20e53e
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>
(cherry picked from commit 63316dcc44518df5e6f5ad53ead05d47aee7d594)