Bug 1688165: Checking if user can view group hompage for embedded images
Rather than checking if user has a group role.
As the embedding of images for group description is linked to the
group homepage, so if you can see that you should be allowed to see
the embedded images within it.

behatnotneeded

Change-Id: I81cbe5cff7229967095a6e2293576861202432c0
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>
(cherry picked from commit 4dc2d4891b0108027851fcb70413af4f24b77685)