Bug 1710710: Placement of public search box vs language selector
When we have $cfg->publicsearchallowed = true;
and are viewing the site when logged out and have langpacks installed
we get an overlap of the language selector and searc box.

When we don't have langpacks installed the search box doesn't align to
the right.

This patch tries to align things up correctly

behatnotneeded

Change-Id: I6f2175af0aa9e164f4c4302f4c1e7012423d1235
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>