Bug 1649433: Resume, replace old fileupload field with filebrowser
Notes from Pramith:
@TODO: file uploader closes popup and does not refresh content
- I fixed this but needs to be done in a nicer way
The problem comes from the filebrowser pieform: when adding a
new file, it call the callback js function of the form.
In other places in mahara, this is fine because the forms
didn't need to set the callback so it could be used by the filebrowser element.
But in this page, the form sets the callback to be compositeSaveCallback

Notes from Dianne:
- Replaced 'attachments' key with 'filebrowser' key in employment and
education history classes. This gets correct filebrowser working.
- Fixed the 'Edit' button in 'Achievements' and 'Education and Employment'
by updating js
- Added employment and education callbacks to compositeSaveCallback

Change-Id: I02f881ddde59e82dcc0e9e973195c557d9925bed