Bug 1708959: Fixing 'submitpages' vs 'submittableto' for group
In webservices / csv upload we deal with the variable 'submitpages' in
relation to if one can submit pages to group. But the group
create/update functions and db table expect 'submittableto'.

The csv upload function converts this as expected but the webservices
didn't.

Have made adjustemets to webservices to suit

behatnotneeded

Change-Id: Ia2a593705a85fec88debed63dd55277c8f1acc37
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>