Bug 1735442: Primary buttons in custom theme need border
Before it was transparent, which was fine when header background
colour was dark. But when it is light, eg white, then we can't see
where the button finishes and the page start.

Have made the border the same colour as text of button.

To test: re-save custom theme for institution with some actual change
so system saves new info

behatnotneeded

Change-Id: I925599d51d1c3446700c5439c1aeffda1383d150
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>
(cherry picked from commit 9280e348f18f2af6080fe5c3c53ec45317144fc7)