Bug 1738898: Elasticsearch not indexing access changes correctly
There is an add_to_queue_access() function that is meant to check what
pages have had their access changed since they were indexed. It was
designed to catch the pages which have a start/stop date.

But that is not quite sufficient.

See bug report for more info

behatnotneeded

Change-Id: I160d074fc709c7e94c9c4cacab8f2c5e6c7e9a86
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>