Bug 1740208: Fetch related view for artefacts we can see
Fix the way we find the related views now that we don't have a 'view'
type to restrict to. We can achieve this via the internal '_id' value
as that is a combination of old type and id values.

Also strip tags except the <span> one from highlighted description code
in case there is bad html/js in there.

behatnotneeded

Change-Id: I962accadb352b52dd72b61844f7fe0bd0d586ce2
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>