Bug 1715299-Changing the admin user to normal user
This should be a user test, not admin as we want to check that a user can add pages and collections.
Admin user is now changed to normal user

One of the failing tests was tinymce_mathslate_plugin.feature.
I was able to replicate this locally twice in 10 runs of the test.
Because of that I'm adding a 1 second wait to that test -rjb

Change-Id: Ie1fd1e08abda501669f7691a8efbf72d031e7fc7