Bug 1763163: Copy group portfolios automatically to existing group members
A portfolio shall be copied automatically to existing regular group 
members, i.e. excluding group tutors and administrators, when
setting up the sharing permissions for the portfolio and selecting the
copying option in "Advanced options".

When new people are added to the group, the portfolio is not automatically
copied to their accounts as they were not in the group when the copying
was turned on.

behatnotneeded

Change-Id: I5f706056e3c53f1edc8bcb86cfd41bad1df0185b