Bug 1763163: record who has been given what page for one time copy
This changes the parent patch a bit. Instead of recording on the view
table if a page is able to be copied to existing group members we
insetad record in a new existing copy table if a group page/collection
has been copied to a member and if not make the copy.

We've also made the switch a slide switch, eg when switching it to
'Yes' it does the task on save and then switches back to 'No'.

This means at a later time if set to 'Yes' again it will do the copying
again and give any newer members a copy of all pages/collections as
well as give older members copies of any newer pages / collections /
collection pages since the previous switching to 'No'

Note: Currently adding new collection pages to older members will only
add the pages as singular pages to their portfolio.

behatnotneeded

Change-Id: Ia22d8138d80ec5fdb5a401de7289b256f9e5fc69
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>