Bug 1770496: Validate access for member to a public group
When site public pages are turned off.

If the user is a member of a group that has 'public' access and the
site 'Allow public pages' are turned off, the user should be able to
view the group's home page - therefore allowing them to paginate
through the "Group portfolios" block in the group home page.

behatnotneeded

Change-Id: I1ea49927532f967c8a676b8a1590d1db3787cfc2
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>