Bug 1771362: Fix timezone issues.
Adds timezone dropdown the site settings. If no timezone selected
it uses site settings country to determine correct timezone and sets
that in the database.
PHP and DB session timezones will then assume this setting for all times.

This makes setting $cfg->dbtimezone obsolete - there is a warning to
admins to update their timezone setting on admin dashboard.

Limitations:
- A timezone or country MUST be selected in site options.

- Will not correct existing incorrect times, as no timezone info was
stored in the DB for dates/times, so it's impossible to determine what timezone
was used when saving.

- Changing the timezone will not adjust the display of existing times, for the
same reason.

behatnotneeded

Change-Id: I34c64cd51ece6758e76dd8c37007dcb316005845