Bug 1782729: Adjust skin form to submit via javascript
In the past there was 2 filebrowser fields, body background and view
background, in the skin edit form and this meant we couldn't submit
the form by js as it confused things.

But now we only have the body background image option so I've altered
the form so that it submits via javascript.

In doing so it gets the changing folder to work as expected as well.

I've also restricted the images one can use to non-profile ones as
that was causing an error on valid image check.

behatnotneeded

Change-Id: Ia11d3ce6073d15d2c60727d3abbdf31b6a93993b
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>