Bug 1790543: Display portfolio instructions
1) Add an 'instructions' field to view table via install / update.

2) Add 'Instructions' textarea / wysiwyg field to view settings page and
have it save to the database

3) Move generic "This area shows a preview of what your page looks like.
Changes are saved automatically. Drag blocks onto the page to add them.
You can then also drag them around the page to change their position."
text in edit page to the help icon and in it's place have an expandable
area to display the saved 'instructions' message.

4) On view display page have an expandable area to display the saved
'instructions' message

5) in view edit mode, collapsing/uncollapsing the Instruction section
is remebered automatically for next time the view is edited

behatnotneeded

Change-Id: Ia56323b9b734b8ebe9498c54dce5a3b934d1fd29