Bug 1781844: Lock blocks on pages
A lot of users in Mahara create (Template) pages to share out to users.
This reduces the setup a new user will have to do to. There is an issue
that on their personal copy, the blocks that were so carefully setup can
be deleted.

This functionality creates a new column in the database called
"lockblocks" that can be enabled/disabled on the share page.

When the user copies the page, this setting is copied over and the
"delete" icon is hidden from view should the lockblocks setting be
enabled. This only happens on normal portfolio pages.

To unset - the user can edit the page settings and
disable the lock blocks feature to enable the deletion of the blocks as
per normal.

behatnotneeded
Change-Id: I7a2062f77431d6371f95d7f55bedcc499dd7caa2