Bug 1784778: Peer assessment block
A new block is created that can only be filled with content from a
person who has the “Peer assessment” role for a portfolio. The block's
content is not visible to the page owner until the peer published it.

The block can be selected by the portfolio author for inclusion on any
page, but it can only be used in conjunction with the peer assessor role.
If no peer assessor is selected for a portfolio, the block stays empty.
It is also available on the site and institution level for inclusion into
portfolio pages there (for the creation of templates).

behatnotneeded: will be added in patch 9063

Change-Id: I6f9945f184b4512c345402e1a73f6fc6f21572a7
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>