Bug 1774106: Upgrade SAML for PHP 7.2
The issue is that assert() with a string arg is deprecated.
The patch that fixes this is currently a release candidate,
which I have installed here to see if the patch works.

Requiring people to use memcached, not memcache
Removed checks for mcrypt, no longer required

@TODO:
Replace RC version with release version when it comes out.

behatnotneeded

Change-Id: I111b8cedeb3847ce585ebac0eb576107542a93a2