Bug 1786722 Behat test: Create page from tags
NOTE: Bug #1787860 Text block title is dropped when creating a page content via tags functionality
Needs to be fixed first before this is run sucessfully.

used css elements in property.php
This is for new functionality in Mahara 18.10, which allows a portfolio
owner to create a page based on tags associated with artefacts and
certain blocks.

Change-Id: Ib49ca44374a70db0f82ab622a22f202b75cabeb6