Bug 1813987: Creating a new page with gristack layout
add a new block, place it anywhere in the grid, edit configuration
and delete it.

It will save it in the db on table block_instance_dimension
but not display it in view or edit mode yet

Also included a small fix in xmldb postgres class generator
to handle reserved words in getAlterFieldSQL function

Failing tests:

- most of them are failing when adding a new block to the page
because they expect to have a modal to choose the position
where to place block. That modal was removed for this patch but
there will be a similar one in patch:
 https://reviews.mahara.org/#/c/9952
and those tests will be fixed there

- a few tests failed when they couldn't find text inside the
blocks because they are not expanded to fit the content yet.
This is added in patch:
 https://reviews.mahara.org/#/c/9986
and they will be fixed there

behatnotneeded

Change-Id: If4521a6315f6e8cc5d88693f536946dace359288