Bug 1803101: peer assessment artefact import
Need to define owner and block when creating a
peer assessment artefact while importing

Create a temp block to assign on artefact creation
that later needs to be replaced by the actual block
when it's available

Draft assessments should not be exported or imported.
For exported portfolios created on older version of
Mahara, there is no way of knowing if they are drafts
at the moment of importing because this information
was not included in the leap2A export xml file.
Not importing draft peer assessments should work
from this patch.

It also adds a fix to support
Postgres and MySQL when a table
column name is a reserved word.

Removing not null restriction for "usr" field in "artefact_peer_assessment"
table:
- The author in peer assessment table should be working as comments table.
- For each comment, there is an entry in 'artefact' table that contains
author and authorname, and another entry in 'artefact_comment_comment'
that doesn't have an author field
because we already have one in 'artefact' table.
- We should have something similar on 'artefact_peer_assessment'.
There is already an author field in 'artefact' table,
then the 'usr' field in 'artefact_peer_assessment' should not to have a
'not null' restriction. If the peer assessment artefact has been imported
and the author is not present in this site, then the 'usr'
should be allowed to be null.

behatnotneeded

Change-Id: I182520159d1fcbc072081c2d976cdc68cde5e19d